refactor: use index for views
